Price changed from decimal to int, code formating fixes

+++ App/Domain/Entities/Product.cs
@@ -5,7 +5,7 @@ public class ProductEntity
 {
     public int Id { get; set; }
     required public string Name { get; set; }
-    required public decimal Price { get; set; }
+    required public int Price { get; set; }
     public List<CartEntity> Carts { get; set; } = null!;
     public List<CheckoutEntity> Checkouts { get; set; } = null!;
